agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tkg-agent"
version = "0.1.0"
description = "Self-hostable agent that turns a Telegram channel into a versioned, password-gated knowledge website."
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
]

[project.optional-dependencies]
backfill = ["telethon>=1.36"]
dev = ["ruff", "pytest"]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
